requirements.txt.txt
python-telegram-bot==21.4yt-dlp
